refactor(vite-app): leva os nodes e links para o CVProvider e remove do useNetwork
