+++ packages/vite-app/src/components/network/NetworkView.tsx
@@ -12,17 +12,15 @@ export type NetworkViewProps = {
 }
 
 export function NetworkView({ width, height }: NetworkViewProps) {
-    const { data, setSelected } = useContext(CVContext);
+    const { nodes, links, setSelected } = useContext(CVContext);
 
     const {
         ref,
-        nodes,
-        links,
         focusOnClick,
         nodeLabel,
         linkLabel,
         linkLabelPosition
-    } = useNetwork({ w: width, h: height, ld: data.data, setSelected });
+    } = useNetwork({ w: width, h: height, setSelected });
 
     return (
         <div className={`${network} m-auto`}>
