+++ packages/vite-app/src/lib/useNetwork.ts
@@ -1,65 +1,15 @@
-import { useCallback, useRef, useMemo } from "react";
+import { useCallback, useRef } from "react";
 import { NodeObject, LinkObject } from "react-force-graph-3d";
 import SpriteText from "three-spritetext";
 
-import { JsonLDType } from "@/types";
-
 export type UseNetworkProps = {
-    ld: JsonLDType | null;
     w?: number;
     h?: number;
     setSelected?: (node: NodeObject | null) => void;
 };
 
-export function useNetwork({ ld, setSelected }: UseNetworkProps) {
-    if (!ld) throw new Error("ld must be defined");
-
+export function useNetwork({ setSelected }: UseNetworkProps) {
     const ref = useRef();
-
-    const { nodes, links } = useMemo(() => {
-        let nodes: NodeObject[] = [];
-        if (ld?.compacted) {
-            nodes = (ld?.compacted["@graph"] as NodeObject[]).map((node) => {
-                return {
-                    "@context": ld?.compacted
-                        ? ld?.compacted["@context"]
-                        : undefined,
-                    ...node,
-                };
-            });
-        }
-        let links: LinkObject[] = [];
-        if (ld?.nquads) {
-            links = (ld?.nquads as unknown as Array<any>).reduce(
-                (acc, node) => {
-                    const foundSubject = nodes?.find(
-                        (n) => n.id === node.subject.value
-                    );
-                    const foundObject = nodes?.find(
-                        (n) => n.id === node.object.value
-                    );
-                    if (foundObject && foundSubject) {
-                        // return only relations between two classes, excluding properties
-                        const link = {
-                            source: node.subject.value,
-                            target: node.object.value,
-                            predicate: node.predicate.value.replace(
-                                "http://schema.org/",
-                                ""
-                            ),
-                            value: 10,
-                            // curvature: Math.random(),
-                            // rotation: Math.random(),
-                        };
-                        acc.push(link);
-                    }
-                    return acc;
-                },
-                []
-            );
-        }
-        return { nodes, links };
-    }, [ld]);
 
     const focusOnClick = useCallback(
         (node: NodeObject) => {
@@ -116,8 +66,6 @@ export function useNetwork({ ld, setSelected }: UseNetworkProps) {
 
     return {
         ref,
-        nodes,
-        links,
         focusOnClick,
         nodeLabel,
         linkLabel,
